Added workflow for CI and stacked all folders in src
1.Added workflow for CI.
2.Stacked all core folders in SRC folders
3.Changed the config folder to contain only config files
4.Added prettier dependency with npm commands for formatting
