+++ .github/workflows/integrate.yml
@@ -1,0 +1,18 @@
+name: Node CI
+
+on: [push, pull_request, pull_request_target]
+
+jobs:
+    test_PR:
+        environment: DevEnv
+        runs-on: ubuntu-latest
+
+        steps:
+            - uses: actions/checkout@v2
+            - uses: actions/setup-node@v2
+              with:
+                  node-version: "14"
+            - run: npm ci
+            - run: npm check-format
+            - run: npm build
+            # - run: npm test
