+++ package.json
@@ -1,49 +1,56 @@
 {
-  "name": "refreshus",
-  "version": "1.0.0",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "dev": "cross-env NODE_ENV=production nodemon server"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^0.21.1",
-    "body-parser": "^1.19.0",
-    "browserify": "^17.0.0",
-    "connect-mongo": "^4.4.1",
-    "cors": "^2.8.5",
-    "cross-env": "^7.0.3",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-handlebars": "^5.3.2",
-    "express-session": "^1.17.2",
-    "hbs": "^4.1.2",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.13.2",
-    "morgan": "^1.10.0",
-    "nanoid": "^3.1.23",
-    "netlify-lambda": "^2.0.11",
-    "node-fetch": "^2.6.1",
-    "nodemailer": "^6.6.2",
-    "nodemailer-mailgun-transport": "^2.1.3",
-    "nodemon": "^2.0.12",
-    "passport": "^0.4.1",
-    "passport-google-oauth20": "^2.0.0",
-    "serverless-http": "^2.7.0"
-  },
-  "engines": {
-    "node": ">=10.0.0"
-  },
-  "devDependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Refresh-Us/RefreshUs.git"
-  },
-  "bugs": {
-    "url": "https://github.com/Refresh-Us/RefreshUs/issues"
-  },
-  "homepage": "https://github.com/Refresh-Us/RefreshUs#readme",
-  "description": ""
+    "name": "refreshus",
+    "version": "1.0.0",
+    "main": "server.js",
+    "scripts": {
+        "start": "node server.js",
+        "dev": "env-cmd -f ./config/dev.env nodemon server",
+        "build": "node server.js",
+        "test": "",
+        "format": "prettier --write \"src/!(views)/**/*.{js,html}\" --config .prettierrc",
+        "check-format": "prettier --check \"src/!(views)/**/*.{js,html}\" --config .prettierrc"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "axios": "^0.21.1",
+        "body-parser": "^1.19.0",
+        "browserify": "^17.0.0",
+        "connect-mongo": "^4.4.1",
+        "cors": "^2.8.5",
+        "cross-env": "^7.0.3",
+        "dotenv": "^10.0.0",
+        "express": "^4.17.1",
+        "express-handlebars": "^5.3.2",
+        "express-session": "^1.17.2",
+        "hbs": "^4.1.2",
+        "jsonwebtoken": "^8.5.1",
+        "mongoose": "^5.13.2",
+        "morgan": "^1.10.0",
+        "nanoid": "^3.1.23",
+        "netlify-lambda": "^2.0.11",
+        "node-fetch": "^2.6.1",
+        "nodemailer": "^6.6.2",
+        "nodemailer-mailgun-transport": "^2.1.3",
+        "nodemon": "^2.0.12",
+        "passport": "^0.4.1",
+        "passport-google-oauth20": "^2.0.0",
+        "serverless-http": "^2.7.0"
+    },
+    "engines": {
+        "node": ">=10.0.0"
+    },
+    "devDependencies": {
+        "env-cmd": "^10.1.0",
+        "prettier": "2.3.2"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Refresh-Us/RefreshUs.git"
+    },
+    "bugs": {
+        "url": "https://github.com/Refresh-Us/RefreshUs/issues"
+    },
+    "homepage": "https://github.com/Refresh-Us/RefreshUs#readme",
+    "description": ""
 }
