+++ src/views/addGame.hbs
@@ -1,0 +1,131 @@
+<style>
+    @import
+    url(https://fonts.googleapis.com/css?family=Roboto:400,300,600,400italic); *
+    { margin: 0; padding: 0; box-sizing: border-box; -webkit-box-sizing:
+    border-box; -moz-box-sizing: border-box; -webkit-font-smoothing:
+    antialiased; -moz-font-smoothing: antialiased; -o-font-smoothing:
+    antialiased; font-smoothing: antialiased; text-rendering:
+    optimizeLegibility; } body { font-family: "Roboto", Helvetica, Arial,
+    sans-serif; font-weight: 100; font-size: 12px; line-height: 30px; color:
+    #777; background-image: url('../public/login-back.jpg'); } .container {
+    max-width: 400px; width: 100%; margin: 0 auto; position: relative; }
+    #contact input[type="text"], #contact input[type="email"], #contact
+    input[type="tel"], #contact input[type="url"], #contact textarea, #contact
+    button[type="submit"] { font: 400 12px/16px "Roboto", Helvetica, Arial,
+    sans-serif; } #contact { background: #F9F9F9; padding: 25px; margin: 150px
+    0; box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.2), 0 5px 5px 0 rgba(0, 0, 0,
+    0.24); } #contact h3 { display: block; font-size: 30px; font-weight: 300;
+    margin-bottom: 10px; } #contact h4 { margin: 5px 0 15px; display: block;
+    font-size: 13px; font-weight: 400; } fieldset { border: medium none
+    !important; margin: 0 0 10px; min-width: 100%; padding: 0; width: 100%; }
+    #contact input[type="text"], #contact input[type="email"], #contact
+    input[type="tel"], #contact input[type="url"], #contact textarea { width:
+    100%; border: 1px solid #ccc; background: #FFF; margin: 0 0 5px; padding:
+    10px; } #contact input[type="text"]:hover, #contact
+    input[type="email"]:hover, #contact input[type="tel"]:hover, #contact
+    input[type="url"]:hover, #contact textarea:hover { -webkit-transition:
+    border-color 0.3s ease-in-out; -moz-transition: border-color 0.3s
+    ease-in-out; transition: border-color 0.3s ease-in-out; border: 1px solid
+    #aaa; } #contact textarea { height: 100px; max-width: 100%; resize: none; }
+    #contact button[type="submit"] { cursor: pointer; width: 100%; border: none;
+    background: #4CAF50; color: #FFF; margin: 0 0 5px; padding: 10px; font-size:
+    15px; } #contact button[type="submit"]:hover { background: #43A047;
+    -webkit-transition: background 0.3s ease-in-out; -moz-transition: background
+    0.3s ease-in-out; transition: background-color 0.3s ease-in-out; } #contact
+    button[type="submit"]:active { box-shadow: inset 0 1px 3px rgba(0, 0, 0,
+    0.5); } .copyright { text-align: center; } #contact input:focus, #contact
+    textarea:focus { outline: 0; border: 1px solid #aaa; }
+    ::-webkit-input-placeholder { color: #888; } :-moz-placeholder { color:
+    #888; } ::-moz-placeholder { color: #888; } :-ms-input-placeholder { color:
+    #888; }
+</style>
+
+<div class="container">
+    <form id="contact" action="/admin/add/game" method="post">
+        <h3>ADD A Game</h3>
+        <h4>Contact us for custom quote</h4>
+        <fieldset>
+            <input
+                placeholder="Game Name"
+                type="text"
+                tabindex="1"
+                name="gameName"
+                required
+                autofocus
+            />
+        </fieldset>
+        <fieldset>
+            <label for="genre">Choose a genre:</label>
+            <select id="genre" name="genre" size="3">
+                <option value="Action">Action</option>
+                <option value="Adventure">Adventures</option>
+                <option value="Shooting">Shootings</option>
+                <option value="Racing">Racing</option>
+                <option value="Sports">Sports</option>
+                <option value="War">War</option>
+            </select>
+        </fieldset>
+
+        <fieldset>
+            <input
+                placeholder="Link"
+                type="url"
+                tabindex="3"
+                name="link"
+                required
+            />
+        </fieldset>
+        <fieldset>
+            <input
+                placeholder="Rating"
+                type="text"
+                tabindex="3"
+                name="rating"
+                required
+            />
+        </fieldset>
+        <fieldset>
+            <input
+                placeholder="Enter Image Cover Link"
+                type="url"
+                tabindex="3"
+                name="coverImage"
+                required
+            />
+        </fieldset>
+        <fieldset>
+            <input
+                placeholder="Enter Title Image Link"
+                type="url"
+                tabindex="3"
+                name="titleImage"
+                required
+            />
+        </fieldset>
+        <fieldset>
+            <textarea
+                placeholder="description"
+                tabindex="5"
+                name="description"
+                required
+            ></textarea>
+        </fieldset>
+        <fieldset>
+            <textarea
+                placeholder="trailer"
+                tabindex="5"
+                name="trailer"
+                required
+            ></textarea>
+        </fieldset>
+
+        <fieldset>
+            <button
+                name="submit"
+                type="submit"
+                id="contact-submit"
+                data-submit="...Sending"
+            >Submit</button>
+        </fieldset>
+    </form>
+</div>
