+++ src/views/favGame.hbs
