+++ src/views/favMovie.hbs
