+++ src/views/suggestions.hbs
